Remove nested routes
